--- search-website-functions-v4/api/Lookup/sample.dat
@@ -1,3 +1,0 @@
-{
-    "name": "Hello"
-}
