--- search-website-functions-v4/api/Suggest/sample.dat
@@ -1,1 +1,0 @@
-{q: "w", top: 5, suggester: "sg"}
