+++ search-website-functions-v4/api/local.settings.json.rename
@@ -2,6 +2,7 @@
   "IsEncrypted": false,
   "Values": {
     "AzureWebJobsStorage": "",
+    "AzureWebJobsFeatureFlags": "EnableWorkerIndexing",
     "FUNCTIONS_WORKER_RUNTIME": "python",
     "SearchApiKey": "YOUR-SEARCH-QUERY-KEY",
     "SearchServiceName": "YOUR-SEARCH-RESOURCE-NAME",
