+++ search-website-functions-v4/api/requirements.txt
@@ -4,7 +4,7 @@
 
 azure-common==1.1.28
 azure-core==1.26.0
-azure-functions==1.12.0
+azure-functions==1.17.0
 azure-search-documents==11.3.0
 certifi==2022.12.7
 chardet==5.0.0
