+++ search-website-functions-v4/api/suggest.sample.dat
@@ -1,0 +1,1 @@
+{"q": "w", "top": 5, suggester: "sg"}
